requirements.txt
torch>=1.4.0
torchvision>=0.5.0
numpy>=1.14.0
tensorboard>=2.3.0
scikit-image==0.12.3
pillow==4.2.1
imageio==2.8.0
tqdm
h5py
matplotlib
